requirements.txt
asgiref==3.9.0
Django==5.2.4
django-nested-admin==4.1.1
ffmpeg==1.4
jalali_core==1.0.0
jdatetime==5.2.0
mysqlclient==2.2.7
pillow==11.3.0
python-monkey-business==1.1.0
sqlparse==0.5.3
tzdata==2025.2
